Extract `TestSetup` from `CoreTestCase`.
This decouples the core configuration from specific test classes.
